+++ texthub/modules/det_heads/__init__.py
@@ -1,4 +1,3 @@
-from .pan_head import PanHead
 from .db_head import DBHead
 from .pse_head import PseHead
 from .pa_head import PanCPPHead