+++ texthub/modules/det_heads/pa_head.py
@@ -180,6 +180,7 @@ def _pa_warpper(kernels: np.ndarray, similarity_vectors: np.ndarray, min_area=2)
     _, cc = cv2.connectedComponents(kernels[0], connectivity=4)
     ##kernel 最小连通区域,从label出发将cc中的像素进行聚类
     label_num, label = cv2.connectedComponents(kernels[1], connectivity=4)
+
     return pa_cpp_f(similarity_vectors, label, cc, label_num, min_area)
 
 
@@ -373,5 +374,45 @@ class PANLoss(nn.Module):
 
         return selected_masks
 
+"""
+pse的py实现方式
+"""
+def get_dis(sv1, sv2):
+    return np.linalg.norm(sv1 - sv2)
+def pa_py(text, similarity_vectors, label, label_values, dis_threshold=0.8):
+    pred = np.zeros(text.shape)
+    queue = Queue(maxsize=0)
+    points = np.array(np.where(label > 0)).transpose((1, 0))
+
+    for point_idx in range(points.shape[0]):
+        y, x = points[point_idx, 0], points[point_idx, 1]
+        label_value = label[y, x]
+        queue.put((y, x, label_value))
+        pred[y, x] = label_value
+    # 计算kernel的值
+    d = {}
+    for i in label_values:
+        kernel_idx = label == i
+        kernel_similarity_vector = similarity_vectors[kernel_idx].mean(0)  # 4
+        d[i] = kernel_similarity_vector
+
+    dx = [-1, 1, 0, 0]
+    dy = [0, 0, -1, 1]
+    kernal = text.copy()
+    while not queue.empty():
+        (y, x, label_value) = queue.get()
+        cur_kernel_sv = d[label_value]
+        for j in range(4):
+            tmpx = x + dx[j]
+            tmpy = y + dy[j]
+            if tmpx < 0 or tmpy >= kernal.shape[0] or tmpy < 0 or tmpx >= kernal.shape[1]:
+                continue
+            if kernal[tmpy, tmpx] == 0 or pred[tmpy, tmpx] > 0:
+                continue
+            if np.linalg.norm(similarity_vectors[tmpy, tmpx] - cur_kernel_sv) >= dis_threshold:
+                continue
+            queue.put((tmpy, tmpx, label_value))
+            pred[tmpy, tmpx] = label_value
+    return pred